[gen] Rename reader module to source
