+++ gen/src/source/idna/idna_mapping_table.rs
@@ -15,7 +15,7 @@ use std::collections::BTreeMap;
 
 use regex::Regex;
 
-use reader::utils::read;
+use source::utils::read;
 
 
 lazy_static! {
