+++ gen/src/source/ucd/derived_core_properties.rs
@@ -13,7 +13,7 @@ use std::char;
 use std::collections::BTreeSet;
 use std::str::FromStr;
 
-use reader::utils::read;
+use source::utils::read;
 
 use super::BINARY_PROPERTIES_REGEX;
