+++ gen/src/source/ucd/derived_normalization_props.rs
@@ -15,7 +15,7 @@ use std::str::FromStr;
 
 use regex::Regex;
 
-use reader::utils::read;
+use source::utils::read;
 
 
 lazy_static! {
