+++ gen/src/source/ucd/unicode_data.rs
@@ -14,7 +14,7 @@ use std::str::FromStr;
 
 use regex::Regex;
 
-use reader::utils::read;
+use source::utils::read;
 
 
 lazy_static! {
