+++ gen/src/writer/ucd/bidi.rs
@@ -14,8 +14,8 @@ use std::collections::{BTreeMap, BTreeSet};
 use std::fmt::Display;
 use std::path::Path;
 
-use reader::ucd::unicode_data::UNICODE_DATA;
-use reader::ucd::readme::UNICODE_VERSION;
+use source::ucd::unicode_data::UNICODE_DATA;
+use source::ucd::readme::UNICODE_VERSION;
 
 use writer::utils::tables::{ToRangeCharSet, ToRangeCharTable};
 use writer::common::emit_unicode_version;
