+++ gen/src/writer/ucd/category.rs
@@ -14,8 +14,8 @@ use std::collections::BTreeMap;
 use std::fmt::Display;
 use std::path::Path;
 
-use reader::ucd::unicode_data::UNICODE_DATA;
-use reader::ucd::readme::UNICODE_VERSION;
+use source::ucd::unicode_data::UNICODE_DATA;
+use source::ucd::readme::UNICODE_VERSION;
 
 use writer::utils::tables::ToRangeCharTable;
 use writer::common::emit_unicode_version;
